Rename device groups table file
