+++ src/db/schemas/index.ts
@@ -10,6 +10,6 @@ export * from "./device-types";
 
 export * from "./device-statuses";
 
-export * from "./device_groups";
+export * from "./device-groups";
 
 export * from "./devices";
